+++ src/dlup1up.f90
@@ -114,14 +114,15 @@
 !>
 !> \ingroup ludecomp
 subroutine dlup1up(m,n,L,ldl,R,ldr,p,u,v,w)
+  use iso_fortran_env
     integer, intent(in) :: m, n, ldl, ldr
     integer, intent(inout) :: p(*)
-    double precision, intent(inout) :: L(ldl,*)
-    double precision, intent(inout) :: R(ldr,*)
-    double precision, intent(in) :: u(*)
-    double precision, intent(in) :: v(*)
-    double precision, intent(out) :: w(*)
-    double precision one,tau,tmp
+    real(real64), intent(inout) :: L(ldl,*)
+    real(real64), intent(inout) :: R(ldr,*)
+    real(real64), intent(in) :: u(*)
+    real(real64), intent(in) :: v(*)
+    real(real64), intent(out) :: w(*)
+    real(real64) one,tau,tmp
     parameter (one = 1d0, tau = 1d-1)
     integer k,info,i,j,itmp
     external xerbla,dcopy,daxpy,dtrsv,dger,dgemv,dswap
