+++ src/sqrder.f90
@@ -97,9 +97,10 @@
 !>
 !> \ingroup qrdecomp
 subroutine sqrder(m,n,Q,ldq,R,ldr,j,w)
+  use iso_fortran_env
     integer, intent(in) :: m, n, j, ldq, ldr
-    real, intent(inout) :: Q(ldq,*), R(ldr,*)
-    real, intent(out) :: w(*)
+    real(real32), intent(inout) :: Q(ldq,*), R(ldr,*)
+    real(real32), intent(out) :: w(*)
     external xerbla,scopy,sqrtv1,sqrot,sqrqh
     integer info,i,k
     ! quick return if possible.
