+++ src/sqrshc.f90
@@ -112,9 +112,10 @@
 !>
 !> \ingroup qrdecomp
 subroutine sqrshc(m,n,k,Q,ldq,R,ldr,i,j,w)
+  use iso_fortran_env
     integer, intent(in) :: m, n, k, ldq, ldr, i, j
-    real, intent(inout) :: Q(ldq,*), R(ldr,*)
-    real, intent(out) :: w(*)
+    real(real32), intent(inout) :: Q(ldq,*), R(ldr,*)
+    real(real32), intent(out) :: w(*)
     external xerbla,scopy,sqrtv1,sqrqh,sqhqr,sqrot
     integer info,jj,kk,l
     ! quick return if possible.
